requirements.txt
pubchempy==1.0.4
Pillow==9.5.0